Large commit - many many changes to eliminate compiler warnings with -Wall - not all of them yet.
git-svn-id: https://svn.code.sf.net/p/freemat/code/trunk@2921 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ FreeMat2/libs/libFreeMat/Array.hpp
@@ -787,13 +787,13 @@ public:
   /**
    * Return the number of rows.
    */
-  inline int rows() const {
+  inline size_t rows() const {
     return getDimensionLength(0);
   }
   /**
    * Return the number of columns.
    */
-  inline int columns() const {
+  inline size_t columns() const {
     return getDimensionLength(1);
   }
   /**
